Added Oculus Initialization
Modified how textures are stored slightly.
Added render target creation to the renderer based on the Oculus SDK
samples.

+++ src/MeshRenderer.h
@@ -43,7 +43,7 @@ void MeshRenderer<T>::Initialize(std::vector<T>& vertices, std::vector<UINT>& in
 	ReleaseCOM(mpVB);
 	ReleaseCOM(mpIB);
 
-	mIndexCount = indices.size();
+	mIndexCount = (int)indices.size();
 
 	//Vertex Buffer
 	D3D11_BUFFER_DESC bd;
